requirements.txt
pandas==2.2.3
Pillow==11.2.1
plotly==5.24.1
streamlit==1.44.1
